cleanup unecessary import, and some comemts
